@@ -1,6 +1,6 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: ["./src/**/*.{html,js,ts}"],
+export default {
+  content: ["./index.html", "./src/**/*.{html,vue,js,ts,jsx,tsx}",],
   theme: {
     extend: {},
   },
